# add warning for integer overflow in the SymbolInstancesArray