<!--
Hello! Thanks for contributing. To help your PR be most easily reviewed, please complete
the following checklist:
-->

first stab at warning language 😁 
the condition is based on the assumption that the max valid 16 bit unsigned integer is 65535

## Checklist

 - [x] briefly describe the changes in this PR
 - [x] write tests for all new functionality
 - [x] [document any changes or additions to public APIs](https://github.com/mapbox/mapbox-gl-js/blob/master/docs/README.md) n/a
 - [x] [post scores for all benchmarks on your branch and `master`](https://github.com/mapbox/mapbox-gl-js/blob/master/bench/README.md#running-benchmarks)
 - [x] [do a quick sanity check on the debug page](https://github.com/mapbox/mapbox-gl-js/blob/master/CONTRIBUTING.md#serving-the-debug-page)
 - [ ] get a PR review :+1: / :-1:

ref #2907 

### this branch
buffer: 1,005 ms
fps: 60 fps
frame-duration: 7.8 ms, 0% > 16ms
query-point: 0.45 ms
query-box: 36.59 ms
geojson-setdata-small: 4 ms
geojson-setdata-large: 107 ms

### master
buffer: 909 ms
fps: 60 fps
frame-duration: 7.7 ms, 0% > 16ms
query-point: 0.41 ms
query-box: 40.46 ms
geojson-setdata-small: 6 ms
geojson-setdata-large: 97 ms