@lucaswoj 2016-08-09
Language looks good to me! 

Add a couple test cases and :shipit: