@mollymerp 2016-08-14
@lucaswoj I implemented your suggestions, but should the warning be actually printing to the console when I run `npm test`? Or is there another way to suppress that just for this test?