@lucaswoj 2016-08-15
> should the warning be actually printing to the console when I run `npm test`?

Ideally not. You should be able to work around this by replacing [`sinon.spy`](http://sinonjs.org/docs/#spies) with [`sinon.stub`](http://sinonjs.org/docs/#stubs).