@mollymerp 2016-08-15
`sinon.stub` implemented.

 🚢  on ✅  ?