@mollymerp 2016-08-16
good call @mourner -- should I reset the value of `SymbolBucket.MAX_QUADS` after the test completes?